Add ability to retrieve an activation's result
Previously, client did not provide a convenient way to
retrieve an activation's result.

It was possible to use Swagger's generated code directly,
but that would lead to duplication of the code already
handled by OWskClient, and was not straightforward.

This commit fixes it by adding a direct way, consistent
with already existing functionality.